% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_subset.R
\name{county_subset}
\alias{county_subset}
\title{Subset pedons by county symbol}
\usage{
county_subset(county, mapunit, land_use, year1, year2)
}
\arguments{
\item{pedons}{Soil Profile Collection (pedons) for analysis}
}
\value{
Soil Profile Collection objects in global environment (one for each county)
}
\description{
Subset pedons by county symbol
}
