% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clay.R
\name{plot_clay}
\alias{plot_clay}
\title{Plot clay percentage by depth for group of pedons}
\usage{
plot_clay(pedons, claymin = 0, claymax = 100)
}
\arguments{
\item{pedons}{Soil Profile Collection (pedons) for analysis}

\item{claymin}{clay axis minimum}

\item{claymax}{clay axis maximum}
}
\value{
list containing one plot for each peiid
}
\description{
Plot clay percentage by depth for group of pedons
}
