% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_tex.R
\name{contrast_tex}
\alias{contrast_tex}
\title{Check for Contrasting Surface/Subsurface Textures}
\usage{
contrast_tex(pedons, tex1, tex2, depth)
}
\arguments{
\item{pedons}{Soil Profile Collection (pedons) for analysis}

\item{tex1}{surface texture classes}

\item{tex2}{subsurface texture classes}

\item{depth}{minimum depth to assess subsurface (cm)}
}
\value{
df of outputs
}
\description{
Check for Contrasting Surface/Subsurface Textures
}
\examples{
pedons <- fetchNASIS(from = "pedons")
tex1 <- c("lvfs", "lfs", "ls", "lcos", "vfs", "fs", "s", "cos")
tex2 <- c("fsl", "vfsl", "l", "sil", "si", "scl", "cl", "sicl", "sc", "sic", "c")
depth <- 25
contrast_tex(pedons, tex1, tex2, depth)
}
