% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapunit_CDL.R
\name{mapunit_CDL}
\alias{mapunit_CDL}
\title{Intersect mapunit geometry with CDL}
\usage{
mapunit_CDL(county, mapunit, land_use, year1, year2)
}
\arguments{
\item{county}{county of interest}

\item{mapunit}{mapunit of interest}

\item{land_use}{land use code (examples)}

\item{state}{state of interest}

\item{year}{year of interest}

\item{duration}{desired duration of cover}
}
\value{
shapefile of intersected geometries
}
\description{
Intersect mapunit geometry with CDL
}
